+++ fuzz/fuzz_targets/fuzz_path.rs
@@ -9,18 +9,8 @@ use svgparser::{path, Tokenize};
 
 fuzz_target!(|data: &[u8]| {
     if let Ok(s) = str::from_utf8(data) {
-        let mut p = path::Tokenizer::from_str(s);
-        loop {
-            match p.parse_next() {
-                Ok(token) => {
-                    if token == path::Token::EndOfStream {
-                        break;
-                    }
-                }
-                Err(e) => {
-                    panic!("{:?}", e);
-                }
-            }
-        }
+        let mut p = path::Tokenizer::from_str(s).tokens();
+        for _ in &mut p {}
+        let _ = p.error().unwrap();
     }
 });