+++ fuzz/fuzz_targets/fuzz_svg.rs
@@ -5,25 +5,10 @@ extern crate svgparser;
 
 use std::str;
 
-use svgparser::{svg, Tokenize, Error};
+use svgparser::{svg, Tokenize};
 
 fuzz_target!(|data: &[u8]| {
     if let Ok(s) = str::from_utf8(data) {
-        let mut p = svg::Tokenizer::from_str(s);
-        loop {
-            match p.parse_next() {
-                Ok(token) => {
-                    if token == svg::Token::EndOfStream {
-                        break;
-                    }
-                }
-                Err(e) => {
-                    match e {
-                        Error::InvalidSvgToken(_) => {}
-                        _ => panic!("{:?}", e),
-                    }
-                }
-            }
-        }
+        for _ in svg::Tokenizer::from_str(s).tokens() {}
     }
 });